pyproject.toml
[tool.poetry]
name = "pythonproject3"
version = "0.1.0"
description = ""
authors = ["macbook <kriukov@pos-credir.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
black = "^24.10.0"
ipython = "^8.29.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
isort = "^5.13.2"
flake8 = "^7.1.1"
coverage = "^7.6.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 119
exclude = 'git'


[tool.isort]
line_length = 119